Relax build verification indirect method arity
In order to complete the proposed IR change for #81
the motivating example
  test/pamela/pending/issue-81-nocolons.pamela
was copied to "current legal" PAMELA syntax in
  test/pamela/pending/issue-81-nocolons2.pamela
in order to create
  test/pamela/pending/IR/issue-81-nocolons2.ir.edn
as the basis for
  test/pamela/pending/IR/issue-81-nocolons.ir.edn

The impact of the change is that arity (signature) match
verification when an indirect method is called will be
deferred to when the root-task is available (i.e. the
entire call graph).

Signed-off-by: Tom Marble <tmarble@info9.net>