Added pseudo fixture for handing stdout/stderr for tests
A problem with adding "negative" tests is that we expect
errors and, previously, these actions would blurt error
messages (typically on stderr).

This fix introduces the #'match-eval-out-err macro which
can be used it tests to:
1. Verify the expected exit code (or arbitrary return value)
2. (Optionally) ensure stdout contains a regex.
3. (Optionally) ensure stderr contains a regex.

In addition to testing for a failure exit code this macro
helps ensure that for both positive and negative test cases
that the output has (at least) some sanity as represented
by the given regex and neither stream pollutes the test
output (which could lead to confusion about the overall
success of tests).

Signed-off-by: Tom Marble <tmarble@info9.net>