Add constant arithmetic to pamela (#196)
I added support for arithmetic in constant expressions containing lvars that can be used only in bounds for the present.
There is a case for adding the constant expressions elsewhere, such as in arguments to method invocation, and to class and lvar constructors. I did not add those at present because I wanted to minimise the changes. Before making constant expressions more widespread, we should decide where we want them to be used, and what other primitives we may want to add beyond +,-,*,/

* Added expressions for use in temporal bounds.
They should arguably be allowed elsewhere too, such as as arguments to constructors, arguments to methods calls and as values to various options like probability, cost, reward etc.
I resisted doing all of that so as to minimise the likelihood of breaking something.  I therefore limited the use, temporarily, to temporal bounds -- to cover my immediate needs.
I will come back to add usage elsewhere later -- and probably add more primitive operations.  So far this commit adds:

(+ 2 4)
(- 6 2)
(* 5 6)
(/ 4 5)

All numbers and thus arithmetic in Pamela is floating point.  Should probably add (ceil arg) (floor arg) and (round arg).

Also updated is the propositions.pamela test which has been expanded to test the use of expressions in temporal bounds.

* Updated the .txt files for all tests because the syntax change caused, inevitably, the AST to change.  None of the EDN files needed to be changes, so the changes did not change the IR in any way.

* Updated rubrics.

* Updated changelog.md and doc/PAMELA.md to indicate the added syntax for +,-,*,/ in bounds expressions.

* Update grammar railroad diagram

Co-authored-by: Dan Cerys <dcerys@dollabs.com>